+++ npm/cmd/convertiptable.go
@@ -3,7 +3,7 @@ package main
 import (
 	"fmt"
 
-	"github.com/Azure/azure-container-networking/npm/pkg/debug/dataplane"
+	"github.com/Azure/azure-container-networking/npm/pkg/dataplane"
 	"github.com/spf13/cobra"
 )
