+++ npm/pkg/dataplane/converter.go
@@ -14,9 +14,9 @@ import (
 
 	"github.com/Azure/azure-container-networking/npm/cache"
 	"github.com/Azure/azure-container-networking/npm/http/api"
-	NPMIPtable "github.com/Azure/azure-container-networking/npm/pkg/debug/dataplane/iptables"
-	"github.com/Azure/azure-container-networking/npm/pkg/debug/dataplane/parse"
-	"github.com/Azure/azure-container-networking/npm/pkg/debug/pb"
+	NPMIPtable "github.com/Azure/azure-container-networking/npm/pkg/dataplane/iptables"
+	"github.com/Azure/azure-container-networking/npm/pkg/dataplane/parse"
+	"github.com/Azure/azure-container-networking/npm/pkg/dataplane/pb"
 	"github.com/Azure/azure-container-networking/npm/util"
 	"google.golang.org/protobuf/encoding/protojson"
 )
