+++ npm/pkg/dataplane/converter_test.go
@@ -4,8 +4,8 @@ import (
 	"reflect"
 	"testing"
 
-	NPMIPtable "github.com/Azure/azure-container-networking/npm/pkg/debug/dataplane/iptables"
-	"github.com/Azure/azure-container-networking/npm/pkg/debug/pb"
+	NPMIPtable "github.com/Azure/azure-container-networking/npm/pkg/dataplane/iptables"
+	"github.com/Azure/azure-container-networking/npm/pkg/dataplane/pb"
 	"github.com/Azure/azure-container-networking/npm/util"
 	"github.com/google/go-cmp/cmp"
 )
