+++ npm/pkg/dataplane/testfiles/npmCacheWithCustomFormat.json
